+++ dipy/align/imaffine.py
@@ -49,7 +49,7 @@ import numpy as np
 import numpy.linalg as npl
 import scipy.ndimage as ndimage
 
-from dipy.align import VerbosityLevels, vector_fields as vf
+from dipy.align import vector_fields as vf
 from dipy.align.imwarp import ScaleSpace, get_direction_and_spacings
 from dipy.align.parzenhist import (
     ParzenJointHistogram,
@@ -60,6 +60,7 @@ from dipy.align.scalespace import IsotropicScaleSpace
 from dipy.core.interpolation import interpolate_scalar_2d, interpolate_scalar_3d
 from dipy.core.optimize import Optimizer
 from dipy.testing.decorators import warning_for_keywords
+from dipy.utils import VerbosityLevels
 
 _interp_options = ["nearest", "linear"]
 _transform_method = {}
