+++ dipy/align/imwarp.py
@@ -8,9 +8,11 @@ from nibabel.streamlines import ArraySequence as Streamlines
 import numpy as np
 import numpy.linalg as npl
 
-from dipy.align import Bunch, VerbosityLevels, floating, vector_fields as vfu
+from dipy.align import vector_fields as vfu
 from dipy.align.scalespace import ScaleSpace
+from dipy.align.utils import floating
 from dipy.testing.decorators import warning_for_keywords
+from dipy.utils import Bunch, VerbosityLevels
 
 RegistrationStages = Bunch(
     INIT_START=0,
