+++ dipy/align/tests/test_crosscorr.py
@@ -1,7 +1,8 @@
 import numpy as np
 from numpy.testing import assert_array_almost_equal
 
-from dipy.align import crosscorr as cc, floating
+from dipy.align import crosscorr as cc
+from dipy.align.utils import floating
 from dipy.testing.decorators import set_random_number_generator
