+++ dipy/align/tests/test_metrics.py
@@ -4,8 +4,8 @@ import numpy as np
 from numpy.testing import assert_array_almost_equal, assert_array_equal, assert_raises
 from scipy import ndimage
 
-from dipy.align import floating
 from dipy.align.metrics import CCMetric, EMMetric, SSDMetric
+from dipy.align.utils import floating
 from dipy.testing.decorators import set_random_number_generator
