+++ dipy/align/utils.py
@@ -1,0 +1,3 @@
+import numpy as np
+
+floating = np.float32
