+++ ruff.toml
@@ -29,7 +29,6 @@ ignore = [
 ]
 
 [lint.extend-per-file-ignores]
-"dipy/align/__init__.py" = ["E402"]
 "dipy/reconst/*.py" = ["B026"]
 "dipy/viz/__init__.py" = ["F401"]
 "doc/examples/**" = ["E402"]
